pyproject.toml
[project]
name = "ml-etiquetas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "authlib>=1.7.0",
    "fastapi>=0.136.1",
    "google-cloud-logging>=3.11.4",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "pymupdf>=1.27.2.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "requests>=2.33.1",
    "tzdata>=2024.1",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
